% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.nphDesign.R
\name{display.nphDesign}
\alias{display.nphDesign}
\title{Graphic Display of Study Design Generated by finalize.nphDesign() Function}
\usage{
display.nphDesign(nphDesign = D0.A18X1.W1, maxT = 50)
}
\arguments{
\item{nphDesign}{nphDesign object generated by finalize.nphDesign()
Function or explore.nphDesign function. In order to use this function,
must set show.setting = "Y" when calling finalize.nphDesign().}

\item{maxT}{Maximum calendar time graphic display. Default 50 months.}
}
\description{
This function visualizes the following information: (1) Survival curves;
(2) Cumulative event curves; (3) Expected average hazard ratio curve;
(4) Non-centrality parameter curves; (5) Projected z boundary curves; (6) Power curves;
(7) Projected p value boundary curve for non-combination tests
}
\examples{

her2p.dist = list(
control = list(dist = "exponential", median = 6.7),
exp = list(HR = 6.7/12.5, delay = 6),
crossover = list(status = "N", HRx = NULL, Tx = NULL)
) 
her2p.accr = list(
  A = 23,
  xi = 2,
  LTFU = 0
)

#recommend side = 1 for superiority design
#alpha takes priority than sf if alpha is provided.
#sf options: "LDOF", "LDPK"
her2p.alphabeta = list(
  sf = list(type = c("LDOF")),
  alpha = c(0.02, 0.03)/2,
  overall.alpha = 0.025,
  beta = 0.1,
  side = 1,
  timing = c(0.75, 1.0)
)

lr = function(s){1}
fh01 = function(s){1-s}
fh11 = function(s){s*(1-s)}

her2p.test = list(IA1 = list(lr), FA=list(lr, fh01))
 
her2p = finalize.nphDesign(T=c(29, 45.5), n=300, r=1, dist=her2p.dist, 
  accr=her2p.accr,alphabeta=her2p.alphabeta, f.ws = her2p.test,
  show.setting="Y")
 
display.nphDesign(nphDesign=her2p, maxT=50)
 
}
