% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.simulation.R
\name{display.simulation}
\alias{display.simulation}
\title{Display of Simulation Results Produced From simulation.nphDesign Function}
\usage{
display.simulation(s.nphDesign)
}
\arguments{
\item{s.nphDesign}{An object produced from simulation.nphDesign Function}
}
\value{
Display of the graphs
}
\description{
This function plots the density curve of z values and power curve by analysis
}
