% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.nphDesign.R
\name{explore.nphDesign}
\alias{explore.nphDesign}
\title{Graphic Exploration of Study Design Setting}
\usage{
explore.nphDesign(
  maxT = 50,
  dist = dist,
  accr = accr,
  size = size,
  alphabeta = alphabeta,
  AHR.method = "geometric.schoenfeld",
  f.ws = list(IA1 = list(lr), FA = list(lr, fh01)),
  show.setting = "Y",
  non.centrality = "Heetal2021",
  s.leg = list(x = 30, y = 1, txt = c("Control", "Exp. Arm")),
  s.lab = list(xlab = "Survival Time (mo)", ylab = "Survival", main =
    "Survival Curve Per Study Design"),
  e.leg = list(txt = c("Control", "Exp. Arm", "Total")),
  e.lab = list(xlab = "Calendar Time (mo)", ylab = "Cumulative Events", main = "")
)
}
\arguments{
\item{maxT}{Maximum calendar time in design considerations. Default 50 months.}

\item{dist}{An object that includes distributions of two arms in the design assumptions}

\item{accr}{Accrual distribution specifications}

\item{size}{Target sample size}

\item{alphabeta}{Type I and type II erros}

\item{AHR.method}{Method of calculating average hazard ratio: "geometric.schoenfeld" or "Kalbfleisch and Prentice"}

\item{f.ws}{Weight functions for statistical tests.}

\item{show.setting}{Option whether to show the recommended study design settings}

\item{s.leg}{Survival plot legend}

\item{s.lab}{Survival plot labels}

\item{e.leg}{Events plot legend}

\item{e.lab}{Events plot labels}
}
\value{
study design
}
\description{
This function visualizes the following information: (1) Survival curves;
(2) Cumulative event curves; (3) Expected average hazard ratio curve;
(4) Fisher's information curves; (5) Non-centrality parameter curves;
(6) Projected z boundary curves; (7) Power curves;
(8) Projected p value boundary curve for non-combination tests;
}
\examples{


her2p.dist = list(
control = list(dist = "exponential", median = 6.7),
exp = list(HR = 6.7/12.5, delay = 6),
crossover = list(status = "N", HRx = NULL, Tx = NULL)
) 
her2p.accr = list(
  A = 23,
  xi = 2,
  LTFU = 0
)
her2p.size = list(
  n = 300,
  r = 1,
  maturity = 0.7
)

#recommend side = 1 for superiority design
#alpha takes priority than sf if alpha is provided.
#sf options: "LDOF", "LDPK"
her2p.alphabeta = list(
  sf = list(type = c("LDOF")),
  alpha = c(0.02, 0.03)/2,
  overall.alpha = 0.025,
  beta = 0.1,
  side = 1,
  timing = c(0.75, 1.0)
)

lr = function(s){1}
fh01 = function(s){1-s}
fh11 = function(s){s*(1-s)}

her2p.test = list(IA1 = list(lr), FA=list(lr, fh01))
 
explore.nphDesign(dist=her2p.dist, accr=her2p.accr, size=her2p.size, 
     alphabeta=her2p.alphabeta, 
     AHR.method="Kalbfleisch and Prentice", f.ws = her2p.test)
 
}
