% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.dataCut.R
\name{f.dataCut}
\alias{f.dataCut}
\title{Generate a dataset by administrative data cut at a prespecified number of target events}
\usage{
f.dataCut(data, targetEvents = 397)
}
\arguments{
\item{data}{Dataset must include variables:
\describe{
\item{enterTime}{The start time for survival calculation, eg, randomization time}
\item{calendarTime}{The time when event/censoring occurred}
\item{survTime}{Survival time for analysis}
\item{cnsr}{Indicator of event(0) or censor (1)}
}}

\item{targetEvents}{The number of target events for analysis.}
}
\value{
A dataframe including variables

OUTPUT data added two variables:
\describe{
\item{-}{calendarCutoff}
\item{-}{survTimeCut}
\item{-}{cnsrCut}
}
}
\description{
Cut data by targetEvents according to calendar time
}
