% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.extract.R
\name{f.extract}
\alias{f.extract}
\title{Extract The Design Parameters From nphDesign Object
Extract the design parameters from nphDesign object generated from
finalize.nphDesign or explore.nphDesign or wlr.power.maxcombo functions. This
is a utility function for running trial simulations and display the nphDesign.}
\usage{
f.extract(nphDesign)
}
\arguments{
\item{nphDesign}{nphDesign object generated from finalize.nphDesign or explore.nphDesign or wlr.power.maxcombo functions.}
}
\value{
An object with design parameters.

nph = D0.A18X1.W1

f.extract(nphDesign=H0.A18X1.W1)

#@export
}
\description{
Extract The Design Parameters From nphDesign Object
Extract the design parameters from nphDesign object generated from
finalize.nphDesign or explore.nphDesign or wlr.power.maxcombo functions. This
is a utility function for running trial simulations and display the nphDesign.
}
