% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.nEachMonth.R
\name{f.nEachMonth}
\alias{f.nEachMonth}
\title{Determine number of subjects randomized for every month according to specified weight function}
\usage{
f.nEachMonth(N = 600, A = 24, w = 2, r = 2)
}
\arguments{
\item{N}{Total number of subjects in a trial}

\item{A}{Duration of accrual period in months}

\item{w}{Weigth parameter in the non-uniform enrollment pattern, culmulative proportion of enrollment = (t/A)^w}

\item{r}{Randomization ratio r:1}
}
\value{
n0 and n1 for number of subjects for each month in two groups (n0: control; n1: experimental)
}
\description{
Determine number of subjects randomized for every month according to specified weight function
}
\examples{
#Enroll 600 patients in 24 months with accrual weight 2 and 2:1 randomization
f.nEachMonth(N=600, A=24, w=2, r=2)
}
