% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.nphDesign.R
\name{finalize.nphDesign}
\alias{finalize.nphDesign}
\title{Finalizing Study Design Based on Given Analysis Calendar Time (DCO)}
\usage{
finalize.nphDesign(
  T = c(29, 45.5),
  n = 300,
  r = 1,
  dist = dist,
  accr = accr,
  alphabeta = alphabeta,
  f.ws = list(IA1 = list(lr), FA = list(lr, fh01)),
  show.setting = "Y",
  non.centrality = "Heetal2021"
)
}
\arguments{
\item{T}{Analysis time (DCO). If there are multiple analyses, then T is a vector.}

\item{dist}{An object that includes distributions of two arms in the design assumptions}

\item{accr}{Accrual distribution specifications}

\item{alphabeta}{Type I and type II errors}

\item{f.ws}{Weight functions for statistical tests.}

\item{show.setting}{Option whether to show the recommended study design settings}

\item{size}{Target sample size}

\item{title}{Annotation of the design option}
}
\value{
study design object
}
\description{
This function calculates the power based on commonly used specifications, based
on wlr.power.maxcombo() function. In practice, a clinical trial is designed with multiple
endpoints (primary and/or secondary). For study planning purpose, the final study
design may need to consider the same analysis time for all endpoints. This function
is based on the fixed calendar times for this purpose.
}
\examples{

her2p.dist = list(
control = list(dist = "exponential", median = 6.7),
exp = list(HR = 6.7/12.5, delay = 6),
crossover = list(status = "N", HRx = NULL, Tx = NULL)
) 
her2p.accr = list(
  A = 23,
  xi = 2,
  LTFU = 0
)

#recommend side = 1 for superiority design
#alpha takes priority than sf if alpha is provided.
#sf options: "LDOF", "LDPK"
her2p.alphabeta = list(
  sf = list(type = c("LDOF")),
  alpha = c(0.02, 0.03)/2,
  overall.alpha = 0.025,
  beta = 0.1,
  side = 1,
  timing = c(0.75, 1.0)
)

lr = function(s){1}
fh01 = function(s){1-s}
fh11 = function(s){s*(1-s)}

her2p.test = list(IA1 = list(lr), FA=list(lr, fh01))
 
finalize.nphDesign(T=c(29, 45.5), n=300, r=1, dist=her2p.dist, accr=her2p.accr,alphabeta=her2p.alphabeta, f.ws = her2p.test)
 
}
