% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.nphDesign.R
\name{summarize.nphDesign}
\alias{summarize.nphDesign}
\title{Summary of Study Design Object nphDesign Produced by finalize.nphDesign or wlr.power.maxcombo function}
\usage{
summarize.nphDesign(
  nphDesign = PFS.design,
  endpoint = "Endpoint",
  caption = "Summary of Statistical Design"
)
}
\arguments{
\item{nphDesign}{nphDesign object generated from finalize.nphDesign or explore.nphDesign or wlr.power.maxcombo functions.}
}
\value{
A nice formatted study design summary ready for formal documentation
}
\description{
Extract the design parameters from nphDesign object generated from
finalize.nphDesign or explore.nphDesign or wlr.power.maxcombo functions. This
is a utility function for running trial simulations and display the nphDesign.
}
