% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlr.cov.R
\name{wlr.cov}
\alias{wlr.cov}
\title{Covariance Matrix of Two Weighted Log-rank Tests (Stratified) At The Same Analysis}
\usage{
wlr.cov(
  time = c(5, 7, 10, 12, 12, 15, 20, 20),
  event = c(1, 0, 0, 1, 1, 0, 1, 1),
  group = c(0, 1, 0, 1, 0, 1, 0, 1),
  strata1 = NULL,
  strata2 = NULL,
  strata3 = NULL,
  rho1 = 0,
  gamma1 = 0,
  tau1 = NULL,
  s.tau1 = 0.5,
  rho2 = 0,
  gamma2 = 1,
  tau2 = NULL,
  s.tau2 = 0.5,
  f.ws1 = NULL,
  f.ws2 = NULL
)
}
\arguments{
\item{time}{Survival time}

\item{event}{Event indicator; 1 = event, 0 = censor}

\item{group}{Treatment group; 1 = experimental group, 0 = control}

\item{strata1}{Stratification variable 1}

\item{strata2}{Stratification variable 2}

\item{strata3}{Stratification variable 3}

\item{rho1}{Parameter for Fleming-Harrington (rho1, gamma1) weighted log-rank test.}

\item{gamma1}{Parameter for Fleming-Harrington (rho1, gamma1) weighted log-rank test.
For log-rank test, set rho1 = gamma1 = 0.}

\item{tau1}{Cut point for stabilized FH test, sFH(rho1, gamma1, tau1); with weight
function defined as w1(t) = s_tilda1^rho1*(1-s_tilda1)^gamma1, where
s_tilda1 = max(s(t), s.tau1) or max(s(t), s(tau1)) if s.tau1 = NULL
tau1 = Inf reduces to regular Fleming-Harrington test(rho1, gamma1)}

\item{s.tau1}{Survival rate cut S(tau1) at t = tau1; default 0.5, ie. cut at median.
s.tau1 = 0 reduces to regular Fleming-Harrington test(rho1, gamma1)}

\item{rho2}{Parameter for Fleming-Harrington (rho2, gamma2) weighted log-rank test.}

\item{gamma2}{Parameter for Fleming-Harrington (rho2, gamma2) weighted log-rank test.
For log-rank test, set rho2 = gamma2 = 0.}

\item{tau2}{Cut point for stabilized FH test, sFH(rho2, gamma2, tau2); with weight
function defined as w2(t) = s_tilda2^rho2*(1-s_tilda2)^gamma2, where
s_tilda2 = max(s(t), s.tau2) or max(s(t), s(tau2)) if s.tau2 = NULL
tau2 = Inf reduces to regular Fleming-Harrington test(rho2, gamma2)}

\item{s.tau2}{Survival rate cut S(tau2) at t = tau2; default 0.5, ie. cut at median.
s.tau2 = 0 reduces to regular Fleming-Harrington test(rho2, gamma2)}

\item{f.ws1}{Self-defined weight function of survival rate.
For example, f.ws1 = function(s){1/max(s, 0.25)}
When f.ws1 or f.ws2 is specified, the weight function takes them as priority.}

\item{f.ws2}{Self-defined weight function of survival rate.
For example, f.ws2 = function(s){1/max(s, 0.25)}
When f.ws1 or f.ws2 is specified, the weight function takes them as priority.}
}
\value{
An object with dataframes below.
\describe{
\item{data}{dataframe with variables: time, event, group, strata1, strata2, strata3.}
\item{uni.event.time}{dataframe with variables}
\itemize{
\item u.eTime: Unique event times;
\item Y0:      Risk set of control arm at each of unique event times;
\item Y1:      Risk set of experimental arm at each of unique event times;
\item Y:       Risk set of pooled data at each of unique event times;
\item dN0:     Event set of control arm at each of unique event times;
\item dN1:     Event set of experimental arm at each of unique event times;
\item dN:      Event set of pooled data at each of unique event times;
\item s:       Survival time of pooled data by KM method;
\item s.til1:  s.tilda1 defined as, s.til = max(s, s.tau1);
\item s.til2:  s.tilda2 defined as, s.til = max(s, s.tau2);
\item w1:      Weight function w1(t) at each of unique event times;
\item w2:      Weight function w2(t) at each of unique event times;
\item V11:     Variance statistic at each of unique event times for w1;
\item V12:     Covariance statistic at each of unique event times for w1 and w2;
\item V22:     Variance statistic at each of unique event times for w2;
\item strata1 Strata 1 value
\item strata1 Strata 2 value
\item strata1 Strata 3 value
}
\item{corr}{Correlation between two weigthed log-rank score statistics U1 and U2, equivalent to the correlation between two normalized weigthed log-rank statistics Z1 and Z2, where Zi = Ui/sqrt(var(Ui))}
\item{cov}{Covariance between two weighted log-rank score statistics, U1 and U2.}
}
}
\description{
This function calculates the covariance matrix of two weighted log-rank tests at the same analysis time.
The two weight functions are specified by stabilized Fleming-Harrington class
with parameters (rho, gamma, tau, s.tau), where tau and s.tau are thresholds for
survival time and survival rates, respectively. Either tau or s.tau can be specified.
tau = Inf or s.tau = 0 reduces to the Fleming-Harrington test (rho, gamma).
User-defined weight functions f.ws1 and f.ws2 can be used as well. For example,
f.ws1 = function(s){s^rho*(1-s)^gamma} is equivalent to Fleming-Harrington (rho, gamma)
test with parameters specified for rho and gamma.
}
\examples{
wlr.cov(time=rexp(100), event=sample(c(0,1), 100, replace = TRUE), group=c(rep(0, 50), rep(1, 50)), rho1=0, gamma1=0, tau1 = NULL, s.tau1=0.5,rho2=0, gamma2=1, tau2 = NULL, s.tau2=0.5,f.ws1=NULL, f.ws2=NULL)
wlr.cov(time=rexp(100), event=sample(c(0,1), 100, replace = TRUE), group=c(rep(0, 50), rep(1, 50)), rho1=0, gamma1=0, tau1 = NULL, s.tau1=0.5,rho2=0, gamma2=1, tau2 = NULL, s.tau2=0.5,f.ws1=NULL, f.ws2=NULL,strata1=sample(c(1,2), 100, replace = TRUE),strata2=sample(c(1,2), 100, replace = TRUE),strata3=sample(c(1,2), 100, replace = TRUE))

}
