% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlr.R
\name{wlr}
\alias{wlr}
\title{Stratified Weighted Log-rank Test}
\usage{
wlr(
  time = c(5, 7, 10, 12, 12, 15, 20, 20),
  event = c(1, 0, 0, 1, 1, 0, 1, 1),
  group = c(0, 1, 0, 1, 0, 1, 0, 1),
  strata1 = NULL,
  strata2 = NULL,
  strata3 = NULL,
  rho = 0,
  gamma = 1,
  tau = NULL,
  s.tau = 0.5,
  f.ws = NULL,
  side = c("two.sided", "one.sided")
)
}
\arguments{
\item{time}{Survival time}

\item{event}{Event indicator; 1 = event, 0 = censor}

\item{group}{Treatment group; 1 = experimental group, 0 = control}

\item{strata1}{Stratification variable 1}

\item{strata2}{Stratification variable 2}

\item{strata3}{Stratification variable 3}

\item{rho}{Parameter for Fleming-Harrington (rho, gamma) weighted log-rank test.}

\item{gamma}{Parameter for Fleming-Harrington (rho, gamma) weighted log-rank test.
For log-rank test, set rho = gamma = 0.}

\item{tau}{Cut point for stabilized FH test, sFH(rho, gamma, tau); with weight
function defined as w(t) = s_tilda^rho*(1-s_tilda)^gamma, where
s_tilda = max(s(t), s.tau) or max(s(t), s(tau)) if s.tau = NULL
tau = Inf reduces to regular Fleming-Harrington test(rho, gamma)}

\item{s.tau}{Survival rate cut S(tau) at t = tau; default 0.5, ie. cut at median.
s.tau = 0 reduces to regular Fleming-Harrington test(rho, gamma)}

\item{f.ws}{Self-defined weight function of survival rate.
For example, f.ws = function(s){1/max(s, 0.25)}
When f.ws is specified, the weight function takes it as priority.}

\item{side}{Type of test. one.sided or two.sided. default = two.sided}
}
\value{
An object with dataframes below.
\describe{
\item{data}{dataframe including the following variables:}
\itemize{
\item time: Survival time
\item event: Event indicator; 0 = censor; 1 = event
\item group: Group indicator; 0 = control; 1 = experimental treatment
}
\item{uni.event.time}{dataframe of unique event times and intermediate statistics
including the following variables:}
\itemize{
\item u.eTime: Unique event times;
\item Y0:      Risk set of control arm at each of unique event times;
\item Y1:      Risk set of experimental arm at each of unique event times;
\item Y:       Risk set of pooled data at each of unique event times;
\item dN0:     Event set of control arm at each of unique event times;
\item dN1:     Event set of experimental arm at each of unique event times;
\item dN:      Event set of pooled data at each of unique event times;
\item s:       Survival time of pooled data by KM method;
\item s.til:   s.tilda defined as, s.til = max(s, s.tau);
\item w:       Weight function w(t) at each of unique event times;
\item U:       Score statistic at each of unique event times;
\item V:       Variance statistic at each of unique event times;
\item z:       Normalized z-statistic at each of unique event times; z = sum(U)/sqrt(sum(V));
}
\item{test.results.strata}{dataframe of intermediate calculation per stratum}
\itemize{
\item rho:     Stabilized Fleming-Harrington test parameter
\item gamma:   Stabilized Fleming-Harrington test parameter
\item tau:     Stabilized Fleming-Harrington test parameter if provided
\item s.tau:   Stabilized Fleming-Harrington test parameter if provided
\item test.side: two.sided or one.sided
\item chisq:   Chi-square statistic, chisq = z^2
\item z:       Z statistic
\item p:       P value per stratum
\item strata1 Strata 1 value
\item strata1 Strata 2 value
\item strata1 Strata 3 value
}
\item{test.results}{dataframe including the following variables:}
\itemize{
\item rho:     Stabilized Fleming-Harrington test parameter
\item gamma:   Stabilized Fleming-Harrington test parameter
\item tau:     Stabilized Fleming-Harrington test parameter if provided
\item s.tau:   Stabilized Fleming-Harrington test parameter if provided
\item test.side: two.sided or one.sided
\item chisq:   Chi-square statistic, chisq = z^2
\item z:       Z statistic
\item p:       P value
}
}
}
\description{
The weight function includes Stabilized Fleming-Harrington class (rho, gamma, tau, s.tau),
and any user-defined weight function. For the stabilized Fleming-Harrington class,
to produce stabilized weighting function from pooled survival curve, specify either
tau or s.tau, which are thresholds in survival time and survival rate, respectively.
The weight function is based on the pooled survival curve within each strata by default.
}
\examples{
wlr(time=rexp(100), event=sample(c(0,1), 100, replace = TRUE), group=c(rep(0, 50), rep(1, 50)), rho=0, gamma=1, tau = NULL, s.tau=0, strata1=sample(c(1,2), 100, replace = TRUE))
wlr(time=rexp(100), event=sample(c(0,1), 100, replace = TRUE), group=c(rep(0, 50), rep(1, 50)), rho=0, gamma=1, tau = NULL, s.tau=0, strata1=sample(c(1,2), 100, replace = TRUE), strata2=sample(c(1,2), 100, replace = TRUE))
wlr(time=rexp(100), event=sample(c(0,1), 100, replace = TRUE), group=c(rep(0, 50), rep(1, 50)), rho=0, gamma=1, tau = NULL, s.tau=0, strata1=sample(c(1,2), 100, replace = TRUE), strata2=sample(c(1,2), 100, replace = TRUE), strata3=sample(c(1,2), 100, replace = TRUE))
wlr(time=rexp(100), event=sample(c(0,1), 100, replace = TRUE), group=c(rep(0, 50), rep(1, 50)), rho=0, gamma=1, tau = NULL, s.tau=0, strata1=sample(c(1,2), 100, replace = TRUE), strata2=sample(c(1,2), 100, replace = TRUE), strata3=sample(c(3,4), 100, replace = TRUE), f.ws=function(s){1/max(s^2, 0.25)})
wlr(time=c(5,7,10,12,12,15,20,20), event=c(1,0,0,1,1,0,1,1), group=c(0,1,0,1,0,1,0,1), rho=0, gamma=1, tau = 3, s.tau=NULL)
wlr(time=c(5,7,10,12,12,15,20,20), event=c(1,0,0,1,1,0,1,1), group=c(0,1,0,1,0,1,0,1), rho=0, gamma=1, tau = 20, s.tau=NULL)
wlr(time=c(5,7,10,12,12,15,20,20), event=c(1,0,0,1,1,0,1,1), group=c(0,1,0,1,0,1,0,1), rho=0, gamma=1, tau = Inf, s.tau=NULL)
wlr(time=c(12,7,10,5,12,15,20,20), event=c(1,0,0,1,1,0,1,1), group=c(1,1,0,0,0,1,0,1), rho=0, gamma=1, tau = 10, s.tau=NULL)
wlr(time=c(12,7,10,5,12,15,20,20), event=c(1,0,0,1,1,0,1,1), group=c(1,1,0,0,0,1,0,1), rho=0, gamma=1, tau = 10, s.tau=0.5, side="one.sided")
wlr(time=c(12,7,10,5,12,15,20,20), event=c(1,0,0,1,1,0,1,1), group=c(1,1,0,0,0,1,0,1), rho=0, gamma=0, tau = 10, s.tau=0.5, side="one.sided")
wlr(time=c(12,7,10,5,12,15,20,20), event=c(1,0,0,1,1,0,1,1), group=c(1,1,0,0,0,1,0,1), rho=0, gamma=0, tau = 10, s.tau=0, side="one.sided")

}
